requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.2.0
cryptography==2.6.1
dj-database-url==0.5.0
Django==2.2
django-cors-headers==2.5.3
django-heroku==0.3.1
django-rest-knox==4.0.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
Flask==1.0.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
Markdown==3.1
MarkupSafe==1.1.1
openapi-spec-validator==0.2.6
pathlib==1.0.1
psycopg2==2.8.1
pycparser==2.19
PyJWT==1.7.1
python-dotenv==0.10.1
pytz==2018.9
PyYAML==5.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
swagger-ui-bundle==0.0.3
urllib3==1.24.1
websocket==0.2.1
websocket-client==0.55.0
websockets==7.0
Werkzeug==0.14.1
whitenoise==4.1.2
